+++ libs/flutter_adaptive_core/lib/components/adaptive_app.dart
@@ -1,4 +1,5 @@
-import 'package:flutter/material.dart';
+import 'package:flutter/widgets.dart';
+import 'package:flutter/material.dart' show ThemeMode;
 import 'package:flutter_adaptive_core/flutter_adaptive_core.dart';
 
 class AdaptiveApp extends AdaptiveWidget {
@@ -7,9 +8,88 @@ class AdaptiveApp extends AdaptiveWidget {
     this.home,
     this.theme,
     this.debugShowCheckedModeBanner = true,
+    this.navigatorKey,
+    Map<String, WidgetBuilder> this.routes = const <String, WidgetBuilder>{},
+    this.initialRoute,
+    this.onGenerateRoute,
+    this.onGenerateInitialRoutes,
+    this.onUnknownRoute,
+    List<NavigatorObserver> this.navigatorObservers =
+        const <NavigatorObserver>[],
+    this.builder,
+    this.title = '',
+    this.onGenerateTitle,
+    this.color,
+    this.locale,
+    this.localizationsDelegates,
+    this.localeListResolutionCallback,
+    this.localeResolutionCallback,
+    this.supportedLocales = const <Locale>[Locale('en', 'US')],
+    this.showPerformanceOverlay = false,
+    this.checkerboardRasterCacheImages = false,
+    this.checkerboardOffscreenLayers = false,
+    this.showSemanticsDebugger = false,
+    this.shortcuts,
+    this.actions,
+    this.restorationScopeId,
+    this.scrollBehavior,
+    this.darkTheme,
+    this.themeMode = ThemeMode.system,
   });
 
   final Widget? home;
   final AdaptiveThemeData? theme;
   final bool debugShowCheckedModeBanner;
+
+  final GlobalKey<NavigatorState>? navigatorKey;
+
+  final Map<String, WidgetBuilder>? routes;
+
+  final String? initialRoute;
+
+  final RouteFactory? onGenerateRoute;
+
+  final InitialRouteListFactory? onGenerateInitialRoutes;
+
+  final RouteFactory? onUnknownRoute;
+
+  final List<NavigatorObserver>? navigatorObservers;
+
+  final TransitionBuilder? builder;
+
+  final String title;
+
+  final GenerateAppTitle? onGenerateTitle;
+
+  final Color? color;
+
+  final Locale? locale;
+
+  final Iterable<LocalizationsDelegate<dynamic>>? localizationsDelegates;
+
+  final LocaleListResolutionCallback? localeListResolutionCallback;
+
+  final LocaleResolutionCallback? localeResolutionCallback;
+
+  final Iterable<Locale> supportedLocales;
+
+  final bool showPerformanceOverlay;
+
+  final bool checkerboardRasterCacheImages;
+
+  final bool checkerboardOffscreenLayers;
+
+  final bool showSemanticsDebugger;
+
+  final Map<ShortcutActivator, Intent>? shortcuts;
+
+  final Map<Type, Action<Intent>>? actions;
+
+  final String? restorationScopeId;
+
+  final ScrollBehavior? scrollBehavior;
+
+  final AdaptiveThemeData? darkTheme;
+
+  final ThemeMode? themeMode;
 }