+++ libs/flutter_adaptive_macos_ui/lib/widget_builders/macosui_app_builder.dart
@@ -5,16 +5,52 @@ import 'package:flutter_adaptive_core/flutter_adaptive_core.dart';
 class MacosUIAdaptiveAppBuilder extends AdaptiveWidgetBuilder<AdaptiveApp> {
   @override
   Widget build(BuildContext context, AdaptiveApp component) {
-    MacosThemeData data = component.theme?.build(context);
+    // ToDo - If new fields added to AdaptiveThemeData, add new fields to here (copyWith)
+    MacosThemeData? data = (component.darkTheme == null)
+        ? null
+        : MacosThemeData.dark()
+            .copyWith(brightness: component.darkTheme?.brightness);
+
+    var shortCuts = (component.shortcuts is Map<LogicalKeySet, Intent>)
+        ? component.shortcuts as Map<LogicalKeySet, Intent>
+        : null;
+
     return MacosApp(
-      // theme: component.theme?.build(context),
-      theme: MacosThemeData.light(),
-      darkTheme: MacosThemeData.dark(),
-      themeMode: data.brightness == Brightness.light
-          ? ThemeMode.light
-          : ThemeMode.dark,
+      theme: component.theme?.build(context),
+      // theme: MacosThemeData.light(),
+      // darkTheme: MacosThemeData.dark(),
+      // themeMode: data?.brightness == Brightness.light
+      //     ? ThemeMode.light
+      //     : ThemeMode.dark,
       home: component.home,
       debugShowCheckedModeBanner: component.debugShowCheckedModeBanner,
+      navigatorKey: component.navigatorKey,
+      routes: component.routes ?? const <String, WidgetBuilder>{},
+      initialRoute: component.initialRoute,
+      onGenerateRoute: component.onGenerateRoute,
+      onGenerateInitialRoutes: component.onGenerateInitialRoutes,
+      onUnknownRoute: component.onUnknownRoute,
+      navigatorObservers:
+          component.navigatorObservers ?? const <NavigatorObserver>[],
+      builder: component.builder,
+      title: component.title,
+      onGenerateTitle: component.onGenerateTitle,
+      color: component.color,
+      locale: component.locale,
+      localizationsDelegates: component.localizationsDelegates,
+      localeListResolutionCallback: component.localeListResolutionCallback,
+      localeResolutionCallback: component.localeResolutionCallback,
+      supportedLocales: component.supportedLocales,
+      showPerformanceOverlay: component.showPerformanceOverlay,
+      checkerboardRasterCacheImages: component.checkerboardRasterCacheImages,
+      checkerboardOffscreenLayers: component.checkerboardOffscreenLayers,
+      showSemanticsDebugger: component.showSemanticsDebugger,
+      shortcuts: shortCuts,
+      actions: component.actions,
+      restorationScopeId: component.restorationScopeId,
+      scrollBehavior: component.scrollBehavior ?? const MacosScrollBehavior(),
+      darkTheme: data,
+      themeMode: component.themeMode,
     );
   }
 }