+++ libs/flutter_adaptive_core/CHANGELOG.md
@@ -9,3 +9,11 @@
 ## 1.0.0-preview003
 
 * Adding selection to AdaptiveListTile
+
+## 1.0.0-preview004
+
+* Adding common fields to AdativeApp widget
+
+## 1.0.0-preview005
+
+* Edited AdaptiveApp and its builders